Refactor hillfort list view to use MVP
