+++ app/src/main/java/org/wit/hillfort/views/login/LoginPresenter.kt
@@ -2,7 +2,7 @@ package org.wit.hillfort.views.login
 
 import org.jetbrains.anko.intentFor
 import org.wit.hillfort.R
-import org.wit.hillfort.activities.HillfortListActivity
+import org.wit.hillfort.views.hillfortList.HillfortListView
 import org.wit.hillfort.main.MainApp
 import org.wit.hillfort.models.hillfort.HillfortJsonStore
 import org.wit.hillfort.models.user.UserModel
@@ -17,7 +17,7 @@ class LoginPresenter(val view: LoginView) {
       if (foundUser != null) {
         app.currentUser = foundUser
         app.hillforts = HillfortJsonStore(view.applicationContext, foundUser.id)
-        view.startActivityForResult(view.intentFor<HillfortListActivity>(), 0)
+        view.startActivityForResult(view.intentFor<HillfortListView>(), 0)
       } else {
         view.showToast(R.string.text_authFailed)
       }
