+++ app/src/main/res/layout/activity_hillfort_list.xml
@@ -6,7 +6,7 @@
   android:layout_height="match_parent"
   android:background="#fff"
   android:orientation="vertical"
-  tools:context="org.wit.hillfort.activities.HillfortListActivity">
+  tools:context="org.wit.hillfort.views.hillfortList.HillfortListView">
 
   <android.support.design.widget.AppBarLayout
     android:layout_width="match_parent"
