+++ app/src/main/res/layout/card_button.xml
@@ -4,7 +4,7 @@
   android:layout_width="match_parent"
   android:layout_height="wrap_content"
   android:orientation="vertical"
-  tools:context="org.wit.hillfort.activities.HillfortListActivity">
+  tools:context="org.wit.hillfort.views.hillfortList.HillfortListView">
 
   <Button
     android:id="@+id/btnCreate"
